refactor passes
