+++ .env.example
@@ -1,5 +1,13 @@
+QBRAID_RUN_REMOTE_TESTS= #True
+QBRAID_API_KEY= #QBRAID_API_KEY
 AWS_ACCESS_KEY_ID= #AWS_ACCESS_KEY_ID
 AWS_SECRET_ACCESS_KEY= #AWS_SECRET_ACCESS_KEY
 QISKIT_IBM_TOKEN= #QISKIT_IBM_TOKEN
-QBRAID_API_KEY= #QBRAID_API_KEY
-QBRAID_RUN_REMOTE_TESTS= #True
+IONQ_API_KEY= #IONQ_API_KEY
+OQC_AUTH_TOKEN= #OQC_AUTH_TOKEN
+AZURE_TENANT_ID= #AZURE_TENANT_ID
+AZURE_CLIENT_ID= #AZURE_CLIENT_ID
+AZURE_CLIENT_SECRET= #AZURE_CLIENT_SECRET
+AZURE_QUANTUM_SUBSCRIPTION_ID= #AZURE_QUANTUM_SUBSCRIPTION_ID
+AZURE_QUANTUM_WORKSPACE_RG= #AZURE_QUANTUM_WORKSPACE_RG
+AZURE_QUANTUM_WORKSPACE_NAME= #AZURE_QUANTUM_WORKSPACE_NAME
