+++ bin/stamp_dev_version.sh
