+++ qbraid/programs/gate_model/qasm3.py
@@ -9,289 +9,24 @@
 # THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
 
 """
-Module defining OpenQasm2Program and OpenQasm3Program classes.
+Module defining OpenQasm3Program class.
 
 """
 from __future__ import annotations
 
 import re
-from typing import Optional, Union
+from typing import Optional
 
 import numpy as np
-from openqasm3 import dumps
-from openqasm3.ast import (
-    BinaryExpression,
-    BitType,
-    BranchingStatement,
-    ClassicalDeclaration,
-    Concatenation,
-    Expression,
-    Identifier,
-    IndexedIdentifier,
-    IntegerLiteral,
-    Program,
-    QuantumBarrier,
-    QuantumGate,
-    QuantumMeasurementStatement,
-    QuantumReset,
-    QubitDeclaration,
-    RangeDefinition,
-    Statement,
-)
+from openqasm3.ast import BitType, ClassicalDeclaration, Program, QubitDeclaration
 from openqasm3.parser import parse
-from qbraid_core._import import LazyLoader
 
-from qbraid.passes.qasm3 import normalize_qasm_gate_params, rebase
-from qbraid.passes.qasm3.compat import declarations_to_qasm2
+from qbraid.passes.qasm import depth, normalize_qasm_gate_params, rebase
+from qbraid.passes.qasm.analyze import expression_value_option
 from qbraid.programs.exceptions import ProgramTypeError
-from qbraid.programs.typer import Qasm2String, Qasm2StringType, Qasm3String, Qasm3StringType
+from qbraid.programs.typer import Qasm3String, Qasm3StringType
 
 from ._model import GateModelProgram
-
-transpiler = LazyLoader("transpiler", globals(), "qbraid.transpiler")
-
-
-def expression_value(expression: Optional[Union[Expression, RangeDefinition]]) -> int:
-    """Return the size of an expression."""
-    if isinstance(expression, IntegerLiteral):
-        return expression.value
-
-    raise ValueError(f"Invalid expression type: {type(expression)}. Expected IntegerLiteral.")
-
-
-def expression_value_option(expression: Optional[Expression]) -> Optional[int]:
-    """Return the size of an expression."""
-    if expression is None:
-        return None
-
-    return expression_value(expression)
-
-
-# pylint: disable-next=too-many-statements
-def depth(
-    qasm_statements: list[Statement], counts: dict[tuple[str, int], int]
-) -> dict[tuple[str, int], int]:
-    """Return the depth of a list of given qasm statements."""
-    qreg_sizes = {}
-    creg_sizes = {}
-    track_measured = {}
-    max_depth = 0
-    # pylint: disable-next=too-many-nested-blocks
-    for statement in qasm_statements:
-        if isinstance(statement, QubitDeclaration):
-            qreg_name = statement.qubit.name
-            qreg_size = expression_value(statement.size)
-            qreg_sizes[qreg_name] = qreg_size
-            continue
-        if isinstance(statement, ClassicalDeclaration) and isinstance(statement.type, BitType):
-            creg_name = statement.identifier.name
-            creg_size: int = expression_value(statement.type.size)
-            creg_sizes[creg_name] = creg_size
-            for i in range(creg_size):
-                track_measured[f"{creg_name}[{i}]"] = 0
-            continue
-        if isinstance(statement, QuantumGate):
-            qubits_involved = set()
-            if all(isinstance(qubit, IndexedIdentifier) for qubit in statement.qubits):
-                for qubit in statement.qubits:
-                    if isinstance(qubit.name, Identifier):
-                        qreg_name = qubit.name.name
-                        if isinstance(qubit.indices[0], list):
-                            expression = qubit.indices[0][0]
-                        qubit_index = expression_value(expression)
-                        counts[(qreg_name, qubit_index)] += 1
-                        qubits_involved.add((qreg_name, qubit_index))
-                max_involved_depth = max(counts[qubit] for qubit in qubits_involved)
-                for qubit in qubits_involved:
-                    counts[qubit] = max_involved_depth
-            else:
-                for qubit in statement.qubits:
-                    qreg_name = str(qubit.name)
-                    for i in range(qreg_sizes[qreg_name]):
-                        counts[(qreg_name, i)] += 1
-            max_depth = max(counts.values())
-        elif isinstance(statement, QuantumReset):
-            if isinstance(statement.qubits, IndexedIdentifier):
-                qreg_name = statement.qubits.name.name
-                if isinstance(statement.qubits.indices[0], list):
-                    expression = statement.qubits.indices[0][0]
-                qubit_index = expression_value(expression)
-                counts[(qreg_name, qubit_index)] += 1
-            else:
-                qreg_name = statement.qubits.name
-                for i in range(qreg_sizes[qreg_name]):
-                    counts[(qreg_name, i)] += 1
-        elif isinstance(statement, QuantumBarrier):
-            for qubit_identifier in statement.qubits:
-                if isinstance(qubit_identifier, (IndexedIdentifier, Identifier)):
-                    qreg_name = str(qubit_identifier.name)
-                    for i in range(qreg_sizes[qreg_name]):
-                        counts[(qreg_name, i)] = max_depth
-        elif isinstance(statement, QuantumMeasurementStatement):
-            qubit = statement.measure.qubit
-            if isinstance(qubit, IndexedIdentifier):
-                qreg_name = qubit.name.name
-                if isinstance(qubit.indices[0], list):
-                    qubit_expr = qubit.indices[0][0]
-                qubit_index = expression_value(qubit_expr)
-                counts[(qreg_name, qubit_index)] += 1
-                max_depth = max(counts.values())
-                if isinstance(statement.target, IndexedIdentifier):
-                    if isinstance(statement.target.indices[0], list):
-                        creg_expr = statement.target.indices[0][0]
-                        creg_index = expression_value(creg_expr)
-                        creg_name = statement.target.name.name
-                        track_measured[(creg_name, creg_index)] = max_depth
-            else:
-                qreg_name = qubit.name
-                for i in range(qreg_sizes[qreg_name]):
-                    counts[(qreg_name, i)] += 1
-                if isinstance(statement.target, Identifier):
-                    creg = str(statement.target.name)
-                max_depth = max(counts.values())
-                for i in range(creg_sizes[creg]):
-                    track_measured[(creg, i)] = max_depth
-        elif isinstance(statement, BranchingStatement) and isinstance(
-            statement.condition, (BinaryExpression, Concatenation)
-        ):
-            expression = statement.condition.lhs
-            if isinstance(expression, (IndexedIdentifier, Identifier)):
-                creg_name = expression.name
-
-                for creg_index in range(creg_sizes[creg_name]):
-                    if (creg_name, creg_index) not in track_measured:
-                        track_measured[(creg_name, creg_index)] = 0
-
-                required_depth = max(
-                    track_measured[(creg_name, creg_index)]
-                    for creg_index in range(creg_sizes[creg_name])
-                )
-                required_depth = max(required_depth, max_depth)
-                for i in range(creg_sizes[creg_name]):
-                    track_measured[(creg_name, i)] = required_depth
-                qubits: set[str] = set()
-                for sub_statement in statement.if_block + statement.else_block:
-                    if isinstance(sub_statement, QuantumGate):
-                        for qubit in sub_statement.qubits:
-                            if isinstance(qubit.name, Identifier):
-                                qreg_name = qubit.name.name
-                            if isinstance(qubit, IndexedIdentifier):
-                                if isinstance(qubit.indices[0], list):
-                                    expression = qubit.indices[0][0]
-                            qubit_index = expression_value(expression)
-                            qubits.add((qreg_name, qubit_index))
-                    elif isinstance(sub_statement, QuantumMeasurementStatement):
-                        if isinstance(sub_statement.measure.qubit.name, Identifier):
-                            qreg_name = sub_statement.measure.qubit.name.name
-                        if isinstance(sub_statement.measure.qubit, IndexedIdentifier):
-                            if isinstance(sub_statement.measure.qubit.indices[0], list):
-                                expression = sub_statement.measure.qubit.indices[0][0]
-                        if isinstance(expression, Expression):
-                            qubit_index = expression_value(expression)
-                            qubits.add((qreg_name, qubit_index))
-
-                for qubit_id in qubits:
-                    counts[qubit_id] = max(required_depth, counts[qubit_id]) + 1
-
-                max_depth = max(counts.values())
-    return counts
-
-
-def has_measurements(program: Union[Program, str]) -> bool:
-    """Check if the program has any measurement operations."""
-    program = parse(program) if isinstance(program, str) else program
-    for statement in program.statements:
-        if isinstance(statement, QuantumMeasurementStatement):
-            return True
-    return False
-
-
-def remove_measurements(program: Union[Program, str]) -> str:
-    """Remove all measurement operations from the program."""
-    program = parse(program) if isinstance(program, str) else program
-    statements = [
-        statement
-        for statement in program.statements
-        if not isinstance(statement, QuantumMeasurementStatement)
-    ]
-    program_out = Program(statements=statements, version=program.version)
-    program_str = dumps(program_out)
-    if float(program.version) == 2.0:
-        program_str = declarations_to_qasm2(program_str)
-    return program_str
-
-
-class OpenQasm2Program(GateModelProgram):
-    """Wrapper class for OpenQASM 2 strings."""
-
-    def __init__(self, program: Qasm2StringType):
-        super().__init__(program)
-        if not isinstance(program, Qasm2String):
-            raise ProgramTypeError(message=f"Expected 'str' object, got '{type(program)}'.")
-
-    def parsed(self) -> Program:
-        """Parse the program string."""
-        return parse(self._program)
-
-    def _get_bits(self, bit_type: str) -> list[str]:
-        """Return the number of qubits or classical bits in the circuit.
-
-        Args:
-            bit_type: either "q" or "c" for qubits or classical bits, respectively.
-
-        Returns:
-            list: A list of qubits or classical bits. Each bit is a tuple of (register name, index)
-        """
-        matches = re.findall(rf"{bit_type}reg (\w+)\[(\d+)\];", self.program)
-
-        result = []
-        for match in matches:
-            var_name = match[0]
-            n = int(match[1])
-            result.extend([(var_name, i) for i in range(n)])
-        return result
-
-    @property
-    def qubits(self) -> list[str]:
-        """Use regex to extract all qreg definitions from the string"""
-        return self._get_bits("q")
-
-    @property
-    def num_clbits(self) -> int:
-        """Return the number of classical bits in the circuit."""
-        return len(self._get_bits("c"))
-
-    @property
-    def depth(self) -> int:
-        """Return the circuit depth (i.e., length of critical path)."""
-        counts = {q: 0 for q in self.qubits}
-        program = self.parsed()
-        counts = depth(program.statements, counts)
-        return max(counts.values())
-
-    def _unitary(self) -> np.ndarray:
-        """Return the unitary of the QASM"""
-        return transpiler.transpile(self.program, "cirq").unitary()
-
-    def remove_idle_qubits(self) -> None:
-        """Checks whether the circuit uses contiguous qubits/indices,
-        and if not, reduces dimension accordingly."""
-        raise NotImplementedError
-
-    def reverse_qubit_order(self) -> None:
-        """Reverses the qubit ordering of a openqasm program."""
-        raise NotImplementedError
-
-    def transform(self, device) -> None:
-        """Transform program to according to device target profile."""
-        if device.id == "quera_qasm_simulator":
-            self._program = remove_measurements(self.program)
-
-        basis_gates = device.profile.get("basis_gates")
-
-        if basis_gates is not None and len(basis_gates) > 0:
-            transformed_qasm = rebase(self.program, basis_gates)
-            self._program = normalize_qasm_gate_params(transformed_qasm)
 
 
 def auto_reparse(func):
