+++ qbraid/transpiler/conversions/qasm2/qasm2_to_cirq.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
 
-from qbraid.passes.qasm2 import flatten_qasm_program
+from qbraid.passes.qasm import unfold_qasm2
 from qbraid.programs.exceptions import QasmError
 from qbraid.transpiler.annotations import weight
 
@@ -44,7 +44,7 @@ def qasm2_to_cirq(qasm: Qasm2StringType) -> cirq.Circuit:
         Cirq circuit representation equivalent to the input QASM string.
     """
     try:
-        qasm = flatten_qasm_program(qasm)
+        qasm = unfold_qasm2(qasm)
         return cirq_qasm_parser.QasmParser().parse(qasm).circuit
     except cirq_qasm_import.QasmException as err:
         raise QasmError from err
