+++ qbraid/transpiler/conversions/qasm2/qasm2_to_ionq.py
@@ -19,8 +19,8 @@ from typing import TYPE_CHECKING, Any
 
 import openqasm3.ast
 
-from qbraid.passes.qasm3.compat import convert_qasm_pi_to_decimal
-from qbraid.programs.gate_model.qasm import OpenQasm2Program
+from qbraid.passes.qasm.compat import convert_qasm_pi_to_decimal
+from qbraid.programs.gate_model.qasm2 import OpenQasm2Program
 from qbraid.transpiler.annotations import weight
 
 if TYPE_CHECKING:
