+++ requirements-dev.txt
@@ -9,7 +9,7 @@ qiskit>=0.44,<1.3
 # transpiler extras
 ply>=3.6
 cirq-ionq>=1.3,<1.5
-qbraid-qir>=0.2.0
+qbraid-qir>=0.2.0,<=0.2.3
 qiskit-qasm3-import>=0.4.2
 pytket-braket>=0.35.1,<0.39.0; python_version >= "3.10"
 bloqade>=0.15.13
