+++ tests/passes/test_qasm3_decompose.py
@@ -22,9 +22,9 @@ from openqasm3 import ast
 from openqasm3.parser import parse
 
 from qbraid.passes.exceptions import CompilationError, QasmDecompositionError
-from qbraid.passes.qasm3.compat import normalize_qasm_gate_params
-from qbraid.passes.qasm3.decompose import assert_gates_in_basis, decompose_qasm3, rebase
-from qbraid.programs.gate_model.qasm import OpenQasm3Program
+from qbraid.passes.qasm.compat import normalize_qasm_gate_params
+from qbraid.passes.qasm.decompose import assert_gates_in_basis, decompose_qasm3, rebase
+from qbraid.programs.gate_model.qasm3 import OpenQasm3Program
 
 
 @pytest.mark.parametrize(
