+++ tox.ini
@@ -23,6 +23,9 @@ pass_env =
     AZURE_QUANTUM_WORKSPACE_RG
     AZURE_QUANTUM_WORKSPACE_NAME
     OQC_AUTH_TOKEN
+    IONQ_API_KEY
+    QISKIT_IBM_TOKEN
+    QBRAID_API_KEY
     QBRAID_RUN_REMOTE_TESTS
 
 [testenv:unit-tests]
@@ -43,28 +46,28 @@ envdir = .tox/linters
 skip_install = true
 deps = isort
 commands =
-    isort . {posargs} qbraid tools tests
+    isort . {posargs} qbraid bin tests
 
 [testenv:pylint]
 envdir = .tox/linters
 skip_install = true
 deps = pylint
 commands =
-    pylint {posargs} qbraid tools tests --disable=C0103,C0414,E0401,R0801,R0902,R0903,R0911,R0912,R0914,R0917,W0212,W0511 --extension-pkg-whitelist=rustworkx
+    pylint {posargs} qbraid bin tests --disable=C0103,C0414,E0401,R0801,R0902,R0903,R0911,R0912,R0914,R0917,W0212,W0511 --extension-pkg-whitelist=rustworkx
 
 [testenv:black]
 envdir = .tox/linters
 skip_install = true
 deps = black
 commands =
-    black qbraid tools tests {posargs}
+    black qbraid bin tests {posargs}
 
 [testenv:headers]
 envdir = .tox/linters
 skip_install = true
 deps = qbraid-cli>=0.8.2
 commands =
-    qbraid admin headers tests tools qbraid --type=gpl {posargs}
+    qbraid admin headers tests bin qbraid --type=gpl {posargs}
 
 [testenv:linters]
 allowlist_externals = qbraid
